Use Long instead of Int

+++ src/main/kotlin/day10/day10.kt
@@ -25,7 +25,7 @@ private fun secondStar(map: Array<Array<Char>>, path: Array<Point>) {
 
     map.forEachIndexed { row, line ->
         line.forEachIndexed { column, _ ->
-            val point = Point(x = column, y = row)
+            val point = Point(x = column.toLong(), y = row.toLong())
             if (!path.contains(point) && path.pointInPolygon(point)) {
                 sum++
             }
@@ -41,7 +41,7 @@ private fun List<String>.toMap() =
 private fun List<String>.findStartPosition(): Point {
     val row = this.indexOfFirst { it.contains('S') }
     val column = this[row].indexOf('S')
-    return Point(x = column, y = row)
+    return Point(x = column.toLong(), y = row.toLong())
 }
 
 private fun findPath(map: Array<Array<Char>>, startPosition: Point): Array<Point> {
@@ -52,7 +52,7 @@ private fun findPath(map: Array<Array<Char>>, startPosition: Point): Array<Point
     )
 
     var nextPosition = map.findInitialDirection(currentPosition)
-    var nextChar = map[nextPosition.x][nextPosition.y]
+    var nextChar = map[nextPosition.x.toInt()][nextPosition.y.toInt()]
 
     while (nextChar != 'S') {
         val xDifference = (currentPosition.x + 1) - (nextPosition.x + 1)
@@ -61,16 +61,16 @@ private fun findPath(map: Array<Array<Char>>, startPosition: Point): Array<Point
         currentPosition = nextPosition
 
         nextPosition = currentPosition + when (nextChar) {
-            'J' -> if (xDifference == 0) Direction.Left else Direction.Up
-            'F' -> if (xDifference == 0) Direction.Right else Direction.Down
-            'L' -> if (xDifference == 0) Direction.Right else Direction.Up
-            '7' -> if (xDifference == 0) Direction.Left else Direction.Down
+            'J' -> if (xDifference == 0L) Direction.Left else Direction.Up
+            'F' -> if (xDifference == 0L) Direction.Right else Direction.Down
+            'L' -> if (xDifference == 0L) Direction.Right else Direction.Up
+            '7' -> if (xDifference == 0L) Direction.Left else Direction.Down
             '-' -> if (xDifference > 0) Direction.Left else Direction.Right
             '|' -> if (yDifference > 0) Direction.Up else Direction.Down
             else -> throw Exception("No path found $nextChar")
         }
         path.add(currentPosition)
-        nextChar = map[nextPosition.y][nextPosition.x]
+        nextChar = map[nextPosition.y.toInt()][nextPosition.x.toInt()]
     }
 
     return path.toTypedArray()
@@ -78,10 +78,10 @@ private fun findPath(map: Array<Array<Char>>, startPosition: Point): Array<Point
 
 private fun Array<Array<Char>>.findInitialDirection(position: Point): Point {
     return position + when {
-        position.x > 0 && this[position.y][position.x - 1] in listOf('F', 'L', '-') -> Direction.Left
-        position.x < this.first().size && this[position.y][position.x + 1] in listOf('7', 'J', '-') -> Direction.Right
-        position.y > 0 && this[position.y - 1][position.x] in listOf('|', 'F', '7') -> Direction.Up
-        position.y < this.size && this[position.y + 1][position.x] in listOf('L', 'J', '|') -> Direction.Down
+        position.x > 0 && this[position.y.toInt()][position.x.toInt() - 1] in listOf('F', 'L', '-') -> Direction.Left
+        position.x < this.first().size && this[position.y.toInt()][position.x.toInt() + 1] in listOf('7', 'J', '-') -> Direction.Right
+        position.y > 0 && this[position.y.toInt() - 1][position.x.toInt()] in listOf('|', 'F', '7') -> Direction.Up
+        position.y < this.size && this[position.y.toInt() + 1][position.x.toInt()] in listOf('L', 'J', '|') -> Direction.Down
         else -> throw Exception("No path found")
     }
 }
